Move landing sections from main config to their own files
Each section uses a markdown file in `content/landing` with its own layout (called section) based off the original fresh theme. All configuration is done in the front matter of the file since that required the least modification. The weight parameter is used to order sections.

Sections are tinted automatically but a custom background can be set with the `sectionStyle` variable.

A sitemap is generated that has links to the various sections but nothing else.

Based off: https://github.com/sethmacleod/dimension

It's not a perfect system but it's good enough. It may pay to create some sort of default theme later.