Update footer
The largest change here is that footer columns are no longer hard coded. 

If a logo is set only a logo is shown, similar to the original theme. There are still options for the title, links and social icons.

The second largest change is that if a string is provided, rather than a map, it is parsed as markdown. *Some* styles have been slightly updated to make this work.